Break out Valuespace definition into a separate, internal module
So that tests can mess with internals, rather than having to construct
state using functions. (This may or may not be a good idea.)
